+++ redis/src/config.rs
@@ -1,6 +1,3 @@
-#[cfg(feature = "config")]
-use ::config_crate::{ConfigError, Environment};
-
 use crate::{Pool, PoolConfig, RedisResult};
 
 /// Configuration object. By enabling the `config` feature you can
@@ -37,17 +34,6 @@ pub struct Config {
 }
 
 impl Config {
-    /// Create configuration from environment variables.
-    #[cfg(feature = "config")]
-    #[deprecated(
-        since = "0.7.0",
-        note = "Please embed this structure in your own config structure and use `config::Config` directly."
-    )]
-    pub fn from_env(prefix: &str) -> Result<Self, ConfigError> {
-        let mut cfg = ::config_crate::Config::new();
-        cfg.merge(Environment::with_prefix(prefix))?;
-        cfg.try_into()
-    }
     /// Create pool using the current configuration
     pub fn create_pool(&self) -> RedisResult<Pool> {
         let url = self.get_url();