+++ src/managed/config.rs
@@ -19,28 +19,6 @@ impl PoolConfig {
             timeouts: Timeouts::default(),
         }
     }
-    /// Create pool config by reading it from the environment.
-    /// ## Example environment
-    /// ```env
-    /// POOL_MAX_SIZE = 1s
-    /// POOL_TIMEOUTS_WAIT = 1s
-    /// POOL_TIMEOUTS_CREATE = 1s
-    /// POOL_TIMEOUTS_RECYCLE = 1s
-    /// ```
-    /// ## Example usage
-    /// ```rust,ignore
-    /// Config::from_env("POOL")
-    /// ```
-    #[deprecated(
-        since = "0.5.2",
-        note = "Please embed this structure in your own config structure and use `config::Config` directly."
-    )]
-    #[cfg(feature = "config")]
-    pub fn from_env(prefix: &str) -> Result<PoolConfig, ::config_crate::ConfigError> {
-        let mut cfg = ::config_crate::Config::new();
-        cfg.merge(::config_crate::Environment::with_prefix(prefix))?;
-        cfg.try_into()
-    }
 }
 
 impl Default for PoolConfig {